+++ changelog.md
@@ -6,7 +6,7 @@
 
 ### Added
 
--   Add `<OrganizationIcon />`
+-   Add `<OrganizationIcon />`, `<SharedWithUserIcon />`
 
 ## [8.0.0-alpha.52] - 2020-05-07